up to 9/9 (final data push)
This is as far as I got with the bbg terminal.
I accidentally overwrote and lost the FX data, so I need to update that
at a later time. This is the 3rd push of essentially the same numbers,
due to me having to fix some links and learning the Desktop app.